feat(datagrid-dropdown-filter-web): removing logic from rendering component
